Switch from Shared Code + conditional compilation to platform-specific overrides of base core.
