+++ src/Auth0.OidcClient.Core/Auth0ClientOptions.cs
