--- src/Auth0.OidcClient.Shared/Auth0.OidcClient.Shared.projitems
@@ -1,15 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>5d8121d2-4ed5-41b9-8b89-8a5ca8e851ba</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Auth0.OidcClient.Shared</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Auth0Client.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Auth0ClientOptions.cs" />
-  </ItemGroup>
-</Project>
