+++ src/Auth0.OidcClient.UWP/Auth0.OidcClient.UWP.csproj
@@ -109,17 +109,17 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Auth0Client.cs" />
     <Compile Include="PlatformWebView.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Properties\Auth0.OidcClient.UWP.rd.xml" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Auth0.OidcClient.Core\Auth0.OidcClient.Core.csproj">
-      <Project>{1F79DB59-F3A8-45B7-972B-ECD15259D1C9}</Project>
+      <Project>{9fba8910-8228-44e7-b6af-4641bfd7a64a}</Project>
       <Name>Auth0.OidcClient.Core</Name>
     </ProjectReference>
   </ItemGroup>
-  <Import Project="..\Auth0.OidcClient.Shared\Auth0.OidcClient.Shared.projitems" Label="Shared" />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
   </PropertyGroup>
