+++ src/Auth0.OidcClient.UWP/Auth0Client.cs
@@ -1,0 +1,24 @@
+﻿using Windows.Security.Authentication.Web;
+using static IdentityModel.OidcClient.OidcClientOptions;
+
+namespace Auth0.OidcClient
+{
+    /// <summary>
+    /// Primary class for performing authentication and authorization operations with Auth0 using the
+    /// underlying <see cref="IdentityModel.OidcClient.OidcClient"/>.
+    /// </summary>
+    public class Auth0Client : Auth0ClientBase
+    {
+        /// <summary>
+        /// Creates a new instance of the Auth0 OIDC Client.
+        /// </summary>
+        /// <param name="options">The <see cref="Auth0ClientOptions"/> specifying the configuration for the Auth0 OIDC Client.</param>
+        public Auth0Client(Auth0ClientOptions options)
+            : base(options, "uwp", AuthorizeResponseMode.FormPost)
+        {
+        }
+
+        /// <inheritdoc />
+        public override string RedirectUri => WebAuthenticationBroker.GetCurrentApplicationCallbackUri().AbsoluteUri;
+    }
+}
