+++ src/Auth0.OidcClient.WinForms/Auth0Client.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Auth0.OidcClient
+{
+    public class Auth0Client : Auth0ClientBase
+    {
+        public Auth0Client(Auth0ClientOptions options)
+            : base(options, "wpf")
+        {
+        }
+    }
+}
