+++ src/Auth0.OidcClient.Xamarin.iOS/Auth0.OidcClient.Xamarin.iOS.csproj
@@ -44,6 +44,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ActivityMediator.cs" />
+    <Compile Include="Auth0Client.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PlatformWebView.cs" />
   </ItemGroup>
@@ -58,6 +59,5 @@
       <Version>2.9.0</Version>
     </PackageReference>
   </ItemGroup>
-  <Import Project="..\Auth0.OidcClient.Shared\Auth0.OidcClient.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\iOS\Xamarin.iOS.CSharp.targets" />
 </Project>
